+++ docs/apache-airflow/core-concepts/params.rst
@@ -22,8 +22,8 @@ Params
 
 Params enable you to provide runtime configuration to tasks. You can configure default Params in your DAG
 code and supply additional Params, or overwrite Param values, at runtime when you trigger a DAG.
-:class:`~airflow.models.param.Param` values are validated with JSON Schema. For scheduled DAG runs,
-default :class:`~airflow.models.param.Param` values are used.
+:class:`~airflow.sdk.definitions.param.Param` values are validated with JSON Schema. For scheduled DAG runs,
+default :class:`~airflow.sdk.definitions.param.Param` values are used.
 
 Also defined Params are used to render a nice UI when triggering manually.
 When you trigger a DAG manually, you can modify its Params before the dagrun starts.
@@ -33,14 +33,14 @@ DAG-level Params
 ----------------
 
 To add Params to a :class:`~airflow.models.dag.DAG`, initialize it with the ``params`` kwarg.
-Use a dictionary that maps Param names to either a :class:`~airflow.models.param.Param` or an object indicating the parameter's default value.
+Use a dictionary that maps Param names to either a :class:`~airflow.sdk.definitions.param.Param` or an object indicating the parameter's default value.
 
 .. code-block::
    :emphasize-lines: 7-10
 
     from airflow import DAG
     from airflow.decorators import task
-    from airflow.models.param import Param
+    from airflow.sdk import Param
 
     with DAG(
         "the_dag",
@@ -127,7 +127,7 @@ You can change this by setting ``render_template_as_native_obj=True`` while init
     ):
 
 
-This way, the :class:`~airflow.models.param.Param`'s type is respected when it's provided to your task:
+This way, the :class:`~airflow.sdk.definitions.param.Param`'s type is respected when it's provided to your task:
 
 .. code-block::
 
@@ -160,7 +160,7 @@ Another way to access your param is via a task's ``context`` kwarg.
 JSON Schema Validation
 ----------------------
 
-:class:`~airflow.models.param.Param` makes use of `JSON Schema <https://json-schema.org/>`_, so you can use the full JSON Schema specifications mentioned at https://json-schema.org/draft/2020-12/json-schema-validation.html to define ``Param`` objects.
+:class:`~airflow.sdk.definitions.param.Param` makes use of `JSON Schema <https://json-schema.org/>`_, so you can use the full JSON Schema specifications mentioned at https://json-schema.org/draft/2020-12/json-schema-validation.html to define ``Param`` objects.
 
 .. code-block::
 
@@ -195,8 +195,8 @@ JSON Schema Validation
     at time of trigger.
 
 .. note::
-    As of now, for security reasons, one can not use :class:`~airflow.models.param.Param` objects derived out of custom classes. We are
-    planning to have a registration system for custom :class:`~airflow.models.param.Param` classes, just like we've for Operator ExtraLinks.
+    As of now, for security reasons, one can not use :class:`~airflow.sdk.definitions.param.Param` objects derived out of custom classes. We are
+    planning to have a registration system for custom :class:`~airflow.sdk.definitions.param.Param` classes, just like we've for Operator ExtraLinks.
 
 Use Params to Provide a Trigger UI Form
 ---------------------------------------
@@ -207,21 +207,21 @@ Use Params to Provide a Trigger UI Form
 This form is provided when a user clicks on the "Trigger DAG" button.
 
 The Trigger UI Form is rendered based on the pre-defined DAG Params. If the DAG has no params defined, the trigger form is skipped.
-The form elements can be defined with the :class:`~airflow.models.param.Param` class and attributes define how a form field is displayed.
+The form elements can be defined with the :class:`~airflow.sdk.definitions.param.Param` class and attributes define how a form field is displayed.
 
 The following features are supported in the Trigger UI Form:
 
-- Direct scalar values (boolean, int, string, lists, dicts) from top-level DAG params are auto-boxed into :class:`~airflow.models.param.Param` objects.
+- Direct scalar values (boolean, int, string, lists, dicts) from top-level DAG params are auto-boxed into :class:`~airflow.sdk.definitions.param.Param` objects.
   From the native Python data type the ``type`` attribute is auto detected. So these simple types render to a corresponding field type.
   The name of the parameter is used as label and no further validation is made, all values are treated as optional.
-- If you use the :class:`~airflow.models.param.Param` class as definition of the parameter value, the following attributes can be added:
+- If you use the :class:`~airflow.sdk.definitions.param.Param` class as definition of the parameter value, the following attributes can be added:
 
-  - The :class:`~airflow.models.param.Param` attribute ``title`` is used to render the form field label of the entry box.
+  - The :class:`~airflow.sdk.definitions.param.Param` attribute ``title`` is used to render the form field label of the entry box.
     If no ``title`` is defined the parameter name/key is used instead.
-  - The :class:`~airflow.models.param.Param` attribute ``description`` is rendered below an entry field as help text in gray color.
+  - The :class:`~airflow.sdk.definitions.param.Param` attribute ``description`` is rendered below an entry field as help text in gray color.
     If you want to provide special formatting or links you need to use the Param attribute
     ``description_md``. See tutorial DAG :ref:`Params UI example DAG <params-ui-tutorial>` for an example.
-  - The :class:`~airflow.models.param.Param` attribute ``type`` influences how a field is rendered. The following types are supported:
+  - The :class:`~airflow.sdk.definitions.param.Param` attribute ``type`` influences how a field is rendered. The following types are supported:
 
       .. list-table::
         :header-rows: 1
