+++ task_sdk/tests/execution_time/conftest.py
@@ -71,6 +71,8 @@ def mocked_parse(spy_agency):
         from airflow.utils import timezone
 
         dag = DAG(dag_id=dag_id, start_date=timezone.datetime(2024, 12, 3))
+        if what.ti_context.dag_run.conf:
+            dag.params = what.ti_context.dag_run.conf  # type: ignore[assignment]
         task.dag = dag
         t = dag.task_dict[task.task_id]
         ti = RuntimeTaskInstance.model_construct(
@@ -120,6 +122,7 @@ def create_runtime_ti(mocked_parse, make_ti_context):
         start_date: str | datetime = "2024-12-01T01:00:00Z",
         run_type: str = "manual",
         try_number: int = 1,
+        conf=None,
         ti_id=None,
     ) -> RuntimeTaskInstance:
         if not ti_id:
@@ -133,6 +136,7 @@ def create_runtime_ti(mocked_parse, make_ti_context):
             data_interval_end=data_interval_end,
             start_date=start_date,
             run_type=run_type,
+            conf=conf,
         )
 
         startup_details = StartupDetails(
