+++ task_sdk/tests/execution_time/test_task_runner.py
@@ -711,6 +711,7 @@ class TestRuntimeTaskInstance:
 
         # Verify the context keys and values
         assert context == {
+            "params": {},
             "var": {
                 "json": VariableAccessor(deserialize_json=True),
                 "value": VariableAccessor(deserialize_json=False),
@@ -751,6 +752,7 @@ class TestRuntimeTaskInstance:
         context = runtime_ti.get_template_context()
 
         assert context == {
+            "params": {},
             "var": {
                 "json": VariableAccessor(deserialize_json=True),
                 "value": VariableAccessor(deserialize_json=False),
@@ -983,6 +985,36 @@ class TestRuntimeTaskInstance:
                 ),
             )
 
+    def test_get_param_from_context(
+        self, mocked_parse, make_ti_context, mock_supervisor_comms, create_runtime_ti
+    ):
+        """Test that a params can be retrieved from context."""
+
+        class CustomOperator(BaseOperator):
+            def execute(self, context):
+                value = context["params"]
+                print("The dag params are", value)
+
+        task = CustomOperator(task_id="print-params")
+        runtime_ti = create_runtime_ti(
+            dag_id="basic_param_dag",
+            task=task,
+            conf={
+                "x": 3,
+                "text": "Hello World!",
+                "flag": False,
+                "a_simple_list": ["one", "two", "three", "actually one value is made per line"],
+            },
+        )
+        run(runtime_ti, log=mock.MagicMock())
+
+        assert runtime_ti.task.dag.params == {
+            "x": 3,
+            "text": "Hello World!",
+            "flag": False,
+            "a_simple_list": ["one", "two", "three", "actually one value is made per line"],
+        }
+
 
 class TestXComAfterTaskExecution:
     @pytest.mark.parametrize(
@@ -1091,4 +1123,86 @@ class TestXComAfterTaskExecution:
 
         assert str(exc_info.value) == (
             f"Returned dictionary keys must be strings when using multiple_outputs, found 2 ({int}) instead"
+        )
+
+
+class TestDagParamRuntime:
+    def test_dag_param_resolves_from_task(self, create_runtime_ti, mock_supervisor_comms, time_machine):
+        """Test dagparam resolves on operator execution"""
+        instant = timezone.datetime(2024, 12, 3, 10, 0)
+        time_machine.move_to(instant, tick=False)
+
+        dag = DAG(dag_id="dag_with_dag_params", start_date=timezone.datetime(2024, 12, 3))
+        dag.param("value", default="NOTSET")
+
+        class CustomOperator(BaseOperator):
+            def execute(self, context):
+                assert dag.params["value"] == "NOTSET"
+
+        task = CustomOperator(task_id="task_with_dag_params")
+        runtime_ti = create_runtime_ti(task=task, dag_id="dag_with_dag_params")
+
+        run(runtime_ti, log=mock.MagicMock())
+
+        mock_supervisor_comms.send_request.assert_called_once_with(
+            msg=SucceedTask(
+                state=TerminalTIState.SUCCESS, end_date=instant, task_outlets=[], outlet_events=[]
+            ),
+            log=mock.ANY,
+        )
+
+    def test_dag_param_dag_overwrite(self, create_runtime_ti, mock_supervisor_comms, time_machine):
+        """Test dag param is overwritten from dagrun config"""
+        instant = timezone.datetime(2024, 12, 3, 10, 0)
+        time_machine.move_to(instant, tick=False)
+
+        dag = DAG(dag_id="dag_with_dag_params_overwrite", start_date=timezone.datetime(2024, 12, 3))
+        dag.param("value", default="NOTSET")
+
+        class CustomOperator(BaseOperator):
+            def execute(self, context):
+                # important to use self.dag here
+                assert self.dag.params["value"] == "new_value"
+
+        # asserting on the default value when not set in dag run
+        assert dag.params["value"] == "NOTSET"
+        task = CustomOperator(task_id="task_with_dag_params_overwrite")
+
+        # we reparse the dag here, and if conf passed, added as params
+        runtime_ti = create_runtime_ti(
+            task=task, dag_id="dag_with_dag_params_overwrite", conf={"value": "new_value"}
+        )
+        run(runtime_ti, log=mock.MagicMock())
+        mock_supervisor_comms.send_request.assert_called_once_with(
+            msg=SucceedTask(
+                state=TerminalTIState.SUCCESS, end_date=instant, task_outlets=[], outlet_events=[]
+            ),
+            log=mock.ANY,
+        )
+
+    def test_dag_param_dag_default(self, create_runtime_ti, mock_supervisor_comms, time_machine):
+        """ "Test dag param is retrieved from default config"""
+        instant = timezone.datetime(2024, 12, 3, 10, 0)
+        time_machine.move_to(instant, tick=False)
+
+        dag = DAG(
+            dag_id="dag_with_dag_params_default",
+            start_date=timezone.datetime(2024, 12, 3),
+            params={"value": "test"},
+        )
+
+        class CustomOperator(BaseOperator):
+            def execute(self, context):
+                assert dag.params["value"] == "test"
+
+        assert dag.params["value"] == "test"
+        task = CustomOperator(task_id="task_with_dag_params_default")
+        runtime_ti = create_runtime_ti(task=task, dag_id="dag_with_dag_params_default")
+
+        run(runtime_ti, log=mock.MagicMock())
+        mock_supervisor_comms.send_request.assert_called_once_with(
+            msg=SucceedTask(
+                state=TerminalTIState.SUCCESS, end_date=instant, task_outlets=[], outlet_events=[]
+            ),
+            log=mock.ANY,
         )
